edit, undo and delete was tested on Postman waiting to be tested with frontend integration...

+++ src/main/java/com/ecommerce/sw2/Models/Services/ActionServiceImp.java
@@ -1,6 +1,7 @@
 package com.ecommerce.sw2.Models.Services;
 
 import com.ecommerce.sw2.Models.Domain.Action;
+import com.ecommerce.sw2.Models.Domain.AddProduct;
 import com.ecommerce.sw2.Models.Domain.Product;
 import com.ecommerce.sw2.Models.Repository.ActionRepository;
 import com.ecommerce.sw2.Models.Repository.ProductBackUpRepository;
@@ -26,10 +27,22 @@ public class ActionServiceImp implements ActionService {
     @Autowired
     StoreRepository storeRepository;
 
+    @Autowired
+    ProductService productService;
+
     @Override
     public Product undo(Long id) {
-        Action action =actionRepository.getOne(id);
+        Action action = actionRepository.getOne(id);
+        if (action == null)
+            return null;
+        if (action.getProductBackup().getMy_id() == null)
+            return null;
+        Long pid = action.getProductBackup().getMy_id();
         Product p = action.Undo(id,actionRepository,productRepository,productBackUpRepository , storeRepository);
+        if (action instanceof AddProduct){
+            productService.MakeProductBackUpPIDWithNulls(pid);
+            productService.MakeProductInCartProductIDWithNulls(pid);
+        }
         actionRepository.delete(action);
         productBackUpRepository.delete(action.getProductBackup());
         return p;
